refact: move gameservice to models
